쿠폰 SAGA 기본 기능 및 변경사항 반영

+++ src/main/kotlin/com/yosep/coupon/coupon/service/CouponByUserCommandService.kt
@@ -2,6 +2,7 @@ package com.yosep.coupon.coupon.service
 
 import com.yosep.coupon.common.data.RandomIdGenerator
 import com.yosep.coupon.common.exception.NotExistElementException
+import com.yosep.coupon.common.exception.NotExistUserException
 import com.yosep.coupon.coupon.data.jpa.dto.CouponByUserCreationDto
 import com.yosep.coupon.coupon.data.jpa.entity.CouponByUser
 import com.yosep.coupon.coupon.data.jpa.entity.CouponState
@@ -34,21 +35,16 @@ class CouponByUserCommandService @Autowired constructor(
             HttpMethod.GET, null, Boolean::class.java)
 
         if(response.body == false) {
-            throw RuntimeException("해당 유저아이디가 존재하지 않습니다.")
+            throw NotExistUserException("해당 유저아이디가 존재하지 않습니다.")
         }
 
         if (optionalCoupon.isEmpty) {
             throw NotExistElementException("쿠폰이 존재하지 않습니다.")
         }
 
-        var id = ""
-        if(couponByUserCreationDto.id == "" || couponByUserCreationDto.id.isEmpty()) {
+        var id = RandomIdGenerator.generate()
+        while (couponByUserRepository.findById(id).isPresent) {
             id = RandomIdGenerator.generate()
-            while (couponByUserRepository.findById(id).isPresent) {
-                id = RandomIdGenerator.generate()
-            }
-        }else {
-            id = couponByUserCreationDto.id
         }
 
         val selectedCoupon = optionalCoupon.get()
@@ -64,7 +60,45 @@ class CouponByUserCommandService @Autowired constructor(
         return couponByUserRepository.save(couponByUser)
     }
 
+    @Transactional(
+        readOnly = false,
+        rollbackFor = [RuntimeException::class, NotExistElementException::class]
+    )
+    @Lock(value = LockModeType.PESSIMISTIC_WRITE)
+    fun ownCouponForTest(couponByUserCreationDto: CouponByUserCreationDto): CouponByUser {
+        val couponId = couponByUserCreationDto.couponId
+        val optionalCoupon = couponRepository.findById(couponId)
+        val response = restTemplate.exchange("http://localhost:8001/users?userId=${couponByUserCreationDto.userId}",
+            HttpMethod.GET, null, Boolean::class.java)
+
+        if(response.body == false) {
+            throw RuntimeException("해당 유저아이디가 존재하지 않습니다.")
+        }
+
+        if (optionalCoupon.isEmpty) {
+            throw NotExistElementException("쿠폰이 존재하지 않습니다.")
+        }
+
+        var id = couponByUserCreationDto.id
+
+        val selectedCoupon = optionalCoupon.get()
+        val couponByUser = CouponByUser(
+            id,
+            couponByUserCreationDto.userId,
+            selectedCoupon,
+            CouponState.READY,
+        )
+
+        selectedCoupon.couponStock.decrease(1)
+
+        return couponByUserRepository.save(couponByUser)
+    }
+
     fun deleteCouponByUser(id: String) {
+        if(couponByUserRepository.findById(id).isEmpty) {
+            return
+        }
+
         couponByUserRepository.deleteById(id)
     }
 }
